+++ src/components/search/SearchList.tsx
@@ -1,30 +1,21 @@
-import { useState } from "react";
 import { Link } from "react-router-dom";
 import SearchPagenation from "./SearchPagenation";
 import { ListMap } from "../shared";
 import styled from "styled-components";
 import { logEvent } from "../../utils/amplitude";
+import usePagenation from "../../hooks/usePagenation";
 
 interface ISearchListProps {
   filteredList: CourseType[] | undefined;
 }
 
 const SearchList = ({ filteredList }: ISearchListProps) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [postsPerPage] = useState(12);
+  const { firstPostIndex, lastPostIndex, pages, currentPage, setCurrentPage } =
+    usePagenation(filteredList, 12);
 
-  const lastPostIndex = currentPage * postsPerPage;
-  const firstPostIndex = lastPostIndex - postsPerPage;
   const currentPosts = filteredList
     ? filteredList.slice(firstPostIndex, lastPostIndex)
     : null;
-
-  const totalContents: any = filteredList?.length;
-
-  const pages = [];
-  for (let i = 1; i <= Math.ceil(totalContents / postsPerPage); i++) {
-    pages.push(i);
-  }
 
   return (
     <div className="my-10 3xl:w-[60%] 2xl:w-[70%] w-[90%] min-h-[1500px]">
@@ -65,7 +56,7 @@ const SearchList = ({ filteredList }: ISearchListProps) => {
             </p>
             <p className="ml-1 mt-2 font-medium  text-gray-400 sm:text-xl mb-3  ">
               {JSON.parse(item.createdAt).substr(0, 10)}{" "}
-              {Number(JSON.parse(item.createdAt).substr(11, 2)) + 9}
+              {Number(JSON.parse(item.createdAt).substr(11, 2)) + 9}:
               {JSON.parse(item.createdAt).substr(14, 2)}
             </p>
           </Link>
